Add seperate package @benzene/jit
