+++ packages/http/package.json
@@ -26,7 +26,7 @@
   "types": "./index.d.ts",
   "exports": {
     ".": {
-      "export": "./index.js",
+      "import": "./index.js",
       "require": "./index.cjs"
     }
   },
